+++ kernel/trace/rethook.c
@@ -202,7 +202,7 @@ void rethook_hook(struct rethook_node *node, struct pt_regs *regs, bool mcount)
 NOKPROBE_SYMBOL(rethook_hook);
 
 /* This assumes the 'tsk' is the current task or is not running. */
-static unsigned long __rethook_find_ret_addr(struct task_struct *tsk,
+static uintptr_t __rethook_find_ret_addr(struct task_struct *tsk,
 					     struct llist_node **cur)
 {
 	struct rethook_node *rh = NULL;
@@ -241,8 +241,8 @@ NOKPROBE_SYMBOL(__rethook_find_ret_addr);
  *
  * Returns found address value or zero if not found.
  */
-unsigned long rethook_find_ret_addr(struct task_struct *tsk, unsigned long frame,
-				    struct llist_node **cur)
+uintptr_t rethook_find_ret_addr(struct task_struct *tsk, uintptr_t frame,
+				 struct llist_node **cur)
 {
 	struct rethook_node *rhn = NULL;
 	unsigned long ret;
@@ -276,8 +276,7 @@ void __weak arch_rethook_fixup_return(struct pt_regs *regs,
 }
 
 /* This function will be called from each arch-defined trampoline. */
-unsigned long rethook_trampoline_handler(struct pt_regs *regs,
-					 unsigned long frame)
+uintptr_t rethook_trampoline_handler(struct pt_regs *regs, uintptr_t frame)
 {
 	struct llist_node *first, *node = NULL;
 	unsigned long correct_ret_addr;
